divide syncdata/operation/list.go
